docs: Document the DeformEffect wrap-mode change
The change from CoglVertexBuffer to CoglPrimitive led to a change of the
default wrap-mode for the pipeline. Since using REPEAT can introduce
artifacts when sampling outside the [ 0, 1 ] texture coordinates range,
and since the default wrap mode was not documented anyway, we want to
keep the change from REPEAT to CLAMP, but it's worth adding an entry in
the release notes.